JudgeService/pyproject.toml
[project]
name = "judgeservice"
version = "0.1.0"
description = "Dockingjudge Judge Service"
authors = [
    {name = "Tapeline",email = "yt.redstone.mail@gmail.com"}
]
license = {text = "GNU GPL v3.0"}
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "faststream[cli] (>=0.5.34,<0.6.0)",
    "dishka (>=1.4.2,<2.0.0)",
    "aiohttp (>=3.11.11,<4.0.0)",
    "pydantic (>=2.10.5,<3.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "aio-pika (>=9.5.4,<10.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
